Move Tooltip into AddObsButton
